// src/lib/kubb-he/types/EncryptedMessage.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type EncryptedMessage = {
  /**
   * @type integer
   */
  chat_id: number;
  /**
   * @type integer
   */
  sender_id: number;
  /**
   * @type string
   */
  sender_name: string;
  /**
   * @type integer
   */
  receiver_id: number;
  /**
   * @type string
   */
  content: string;
  /**
   * @type string | undefined
   */
  iv?: string;
  /**
   * @type string | undefined
   */
  image?: string;
  /**
   * @type string | undefined
   */
  image_to_classify?: string;
  /**
   * @type string
   */
  type: string;
  /**
   * @type boolean | undefined
   */
  is_typing?: boolean;
  /**
   * @type string, date-time
   */
  timestamp: string;
  /**
   * @type string | undefined
   */
  classification_result?: string;
};